+++ common/src/main/java/net/ixdarklord/coolcat_lib/common/item/ComponentItem.java
@@ -4,6 +4,7 @@ import net.ixdarklord.coolcat_lib.platform.Services;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
@@ -25,21 +26,28 @@ public class ComponentItem extends Item {
         this.componentType = componentType;
     }
 
-    public int getSplitterLength() {
-        return Math.max(200, itemID.toString().length());
-    }
-
     @Override
     public void appendHoverText(@NotNull ItemStack stack, @Nullable Level level, @NotNull List<Component> tooltipComponents, @NotNull TooltipFlag isAdvanced) {
         if (componentType.get() != null) {
-            tooltipComponents.add(new TextComponent("§8| ").append(componentType.get()));
+            if (appendToName()) {
+                MutableComponent name = tooltipComponents.get(0).copy().append(new TextComponent(" | ").withStyle(ChatFormatting.DARK_GRAY).append(componentType.get()));
+                tooltipComponents.set(0, name);
+            } else tooltipComponents.add(new TextComponent("| ").withStyle(ChatFormatting.DARK_GRAY).append(componentType.get()));
         }
+    }
+
+    public boolean appendToName() {
+        return false;
+    }
+
+    public int getSplitterLength() {
+        return Math.max(200, itemID.toString().length());
     }
 
     public boolean isShiftButtonNotPressed(@Nullable List<Component> tooltipComponents) {
         if (!Screen.hasShiftDown()) {
             if (tooltipComponents != null)
-                tooltipComponents.add(new TextComponent("§8➤ ").append(new TranslatableComponent("tooltip.coolcat_lib.press.shift").withStyle(ChatFormatting.GRAY)));
+                tooltipComponents.add(new TextComponent("➤ ").withStyle(ChatFormatting.DARK_GRAY).append(new TranslatableComponent("tooltip.coolcat_lib.press.shift").withStyle(ChatFormatting.GRAY)));
             return true;
         }
         return false;