+++ common/src/main/java/net/ixdarklord/coolcat_lib/platform/services/IPlatformHelper.java
@@ -3,8 +3,11 @@ package net.ixdarklord.coolcat_lib.platform.services;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 
+import java.nio.file.Path;
+
 public interface IPlatformHelper {
     String getPlatformName();
+    Path getConfigDir();
     boolean isModLoaded(String modId);
     boolean isDevelopmentEnvironment();
     void registerConfig();