+++ forge/src/main/java/net/ixdarklord/coolcat_lib/platform/forge/ForgePlatformHelper.java
@@ -5,20 +5,28 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.loading.FMLLoader;
+import net.minecraftforge.fml.loading.FMLPaths;
 import net.minecraftforge.registries.ForgeRegistries;
+
+import java.nio.file.Path;
 
 public class ForgePlatformHelper implements IPlatformHelper {
     @Override
     public String getPlatformName() {
         return "Forge";
     }
+
+    @Override
+    public Path getConfigDir() {
+        return FMLPaths.CONFIGDIR.get();
+    }
+
     @Override
     public boolean isModLoaded(String modId) {
         return ModList.get() != null && ModList.get().isLoaded(modId);
     }
     @Override
     public boolean isDevelopmentEnvironment() {
-        //noinspection UnstableApiUsage
         return !FMLLoader.isProduction();
     }
     @Override