+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 This file is for listing all the changes to this project
 
+## v1.0.2 Release | Dec 27, 2023
+### Adding
+- [Fabric] Conditional Recipes
+- Improving the TomlConfigReader class
+
 ## v1.0.1 Release | Nov 8, 2023
 ### Adding
 - Adding Arabic Translation